Fix API index heading text
