+++ src/api/ApiIndex.vue
@@ -46,23 +46,6 @@ const filtered = computed(() => {
     })
     .filter((i) => i) as APIGroup[]
 })
-
-// same as vitepress' slugify logic
-function slugify(text: string): string {
-  return (
-    text
-      // Replace special characters
-      .replace(/[\s~`!@#$%^&*()\-_+=[\]{}|\\;:"'<>,.?/]+/g, '-')
-      // Remove continuous separators
-      .replace(/\-{2,}/g, '-')
-      // Remove prefixing and trailing separators
-      .replace(/^\-+|\-+$/g, '')
-      // ensure it doesn't start with a number (#121)
-      .replace(/^(\d)/, '_$1')
-      // lowercase
-      .toLowerCase()
-  )
-}
 </script>
 
 <template>
@@ -85,7 +68,7 @@ function slugify(text: string): string {
       :key="section.text"
       class="api-section"
     >
-      <h2 :id="slugify(section.text)">{{ section.text }}</h2>
+      <h2 :id="section.anchor">{{ section.text }}</h2>
       <div class="api-groups">
         <div
           v-for="item of section.items"
@@ -95,7 +78,7 @@ function slugify(text: string): string {
           <h3>{{ item.text }}</h3>
           <ul>
             <li v-for="h of item.headers" :key="h.anchor">
-              <a :href="item.link + '.html#' + slugify(h.anchor)">{{ h.anchor }}</a>
+              <a :href="item.link + '.html#' + h.anchor">{{ h.text }}</a>
             </li>
           </ul>
         </div>
